Policies

+++ app/controllers/submissions/hypotheses_controller.rb
@@ -2,6 +2,7 @@ class Submissions::HypothesesController < ApplicationController
   def create
     @model = Current.user.models.find(params[:submission_id])
     @hypothesis = @model.hypothesis.new(hypothesis_params)
+    authorize(@hypothesis)
 
     if @hypothesis.save
       flash.now[:notice] = "Hypothesis succesfully created"
@@ -13,6 +14,7 @@ class Submissions::HypothesesController < ApplicationController
 
   def destroy
     @hypothesis = Hypothesis.owned_by(Current.user).find(params[:id])
+    authorize(@hypothesis)
 
     if @hypothesis.destroy
       @empty_hypothesis = Hypothesis::Empty.new(@hypothesis.model, @hypothesis.test_set_entry)
